% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_diffexpMat.R
\name{build_diffexpMat}
\alias{build_diffexpMat}
\title{This function is to build a differentially expressed matrix}
\usage{
build_diffexpMat(
  sc.eset,
  cellType,
  only.pos = TRUE,
  test.use = "wilcox",
  logfc.threshold = 0.25,
  min.pct = 0.05,
  return.thresh = 0.01,
  filepath = NULL,
  verbose = TRUE,
  ...
)
}
\arguments{
\item{sc.eset}{SeuratObject for scRNA-seq data}

\item{cellType}{character, name of metadata column indicating cell type}

\item{only.pos}{only return positive gene markers (TRUE by default)}

\item{test.use}{which statistical test to use}

\item{logfc.threshold}{limit testing to genes which show, on average, at least X-fold difference (log-scale) between the two groups of cells}

\item{min.pct}{only test genes that are detected in a minimum fraction of min.pct cells in either of the two populations}

\item{return.thresh}{only return gene markers that have a p-value < return.thresh}

\item{filepath}{directory where the results will be saved, default is NULL}
}
\value{
a list with three attributes:
\item{ctDEGs}{data frame, matrix containing a ranked list of differentially expressed genes, and associated statistics}
\item{diffexpMat}{numeric matrix, a differential expressed matrix}
\item{designMat}{numeric matrix, a gene by cell-type matrix (genome-wide)}
}
\description{
This function is to build a differentially expressed matrix
}
\author{
Wenjun Shen
}
