% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_sigMat.R
\name{build_sigMat}
\alias{build_sigMat}
\title{This function is to build a signature matrix by selecting the top N markers for each cell type}
\usage{
build_sigMat(
  diffexpMat,
  ctDEGs,
  num.top,
  sort.by = "p.value",
  rm.dupli = TRUE,
  filepath = NULL
)
}
\arguments{
\item{diffexpMat}{numeric matrix, a differentially expressed matrix}

\item{ctDEGs}{data frame, matrix containing a ranked list of differentially expressed genes, and associated statistics}

\item{num.top}{the number of top markers}

\item{sort.by}{rank the markers in ascending order by their p-values (p.value, default) or in descending order by their fold changes (fold.change)}

\item{rm.dupli}{whether to exclude markers shared between two or more cell types, default is TRUE}

\item{filepath}{directory where the results will be saved, default is NULL}
}
\value{
a list with two attributes:
\item{topSigMat}{numeric matrix, signature matrix}
\item{topMarkers}{a list of marker genes for each cell type}
}
\description{
This function is to build a signature matrix by selecting the top N markers for each cell type
}
\author{
Wenjun Shen
}
