% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_sigMat_condNum.R
\name{build_sigMat_condNum}
\alias{build_sigMat_condNum}
\title{This function is to build a signature matrix by minimizing the condition number}
\usage{
build_sigMat_condNum(
  diffexpMat,
  ctDEGs,
  increment = seq(50, 300, 50),
  sort.by = "p.value",
  rm.dupli = TRUE,
  filepath = NULL
)
}
\arguments{
\item{diffexpMat}{numeric matrix, a differentially expressed matrix}

\item{ctDEGs}{data frame, matrix containing a ranked list of differentially expressed genes, and associated statistics}

\item{increment}{numeric vector, create multiple candidate matrices by specifying the number of top markers from each cell type}

\item{sort.by}{rank the markers in ascending order by their p-values (p.value, default) or in descending order by their fold changes (fold.change)}

\item{rm.dupli}{whether to exclude markers shared between two or more cell types, default is TRUE}

\item{filepath}{directory where the results will be saved, default is NULL}
}
\value{
a list with four attributes:
\item{sigMat}{numeric matrix, signature matrix}
\item{markers}{a list of marker genes for each cell type}
\item{min.cond.num}{the lowest condiction number}
\item{num.top}{the number of top markers with the lowest condiction number}
}
\description{
This function is to build a signature matrix by minimizing the condition number
}
\author{
Wenjun Shen
}
