% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cal_cellSize.R
\name{cal_cellSize}
\alias{cal_cellSize}
\title{This function is to estimate the cell sizes by using ERCC spike-in controls}
\usage{
cal_cellSize(sc.eset, cellType, probs = 0.75)
}
\arguments{
\item{sc.eset}{SeuratObject for scRNA-seq data}

\item{cellType}{character, name of metadata column indicating cell type}

\item{probs}{numeric vector of probabilities with values in [0, 1], default is 0.75}
}
\value{
cellSize a vector of calculated cell sizes
}
\description{
This function is to estimate the cell sizes by using ERCC spike-in controls
}
\author{
Wenjun Shen
}
