% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CSsingle_basis.R
\name{CSsingle_basis}
\alias{CSsingle_basis}
\title{This function is to estimate the cell-type proportions from the individual bulk mixture \code{y}}
\usage{
CSsingle_basis(X, y, markers, cellSize = NULL, dampened = FALSE)
}
\arguments{
\item{X}{numeric matrix, signature matrix}

\item{y}{numeric vector of bulk expression}

\item{markers}{list of gene markers for each cell type}

\item{cellSize}{Users can specify a numeric vector of cell sizes, default is NULL, which means that we assume the absolute amount of total mRNA is similar across different cell types in deconvolution}

\item{dampened}{whether to introduce a upper bound constant that limits the maximum value any weight can take on, default is FALSE}
}
\value{
a numeric vector of cell type proportions (add up to 1)
}
\description{
This function is to estimate the cell-type proportions from the individual bulk mixture \code{y}
}
\author{
Wenjun Shen
}
