% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CSsingle_optimize.R
\name{CSsingle_optimize}
\alias{CSsingle_optimize}
\title{This function is to combine CSsingle with each signature matrix to get estimates of the cell type proportions. The signature matrices were ordered in descending order based on the Spearman's correlation between their inferred and real bulk gene expression data.}
\usage{
CSsingle_optimize(
  diffexpMat,
  mixture,
  ctDEGs,
  cellSize = NULL,
  dampened = FALSE,
  increment = seq(50, 300, 50),
  sort.by = "p.value",
  rm.dupli = TRUE,
  filepath = NULL
)
}
\arguments{
\item{diffexpMat}{numeric matrix, a differentially expressed matrix}

\item{mixture}{bulk gene expression matrix}

\item{ctDEGs}{data frame, matrix containing a ranked list of differentially expressed genes, and associated statistics}

\item{cellSize}{Users can specify a numeric vector of cell sizes, default is NULL, which means that we assume the absolute amount of total mRNA is similar across different cell types in deconvolution}

\item{dampened}{whether to introduce a upper bound constant that limits the maximum value any weight can take on, default is FALSE}

\item{increment}{numeric vector, create multiple candidate matrices by specifying the number of top markers from each cell type}

\item{sort.by}{rank the markers in ascending order by their p-values (p.value, default) or in descending order by their fold changes (fold.change)}

\item{rm.dupli}{whether to exclude markers shared between two or more cell types, default is TRUE}

\item{filepath}{directory where the results will be saved, default is NULL}
}
\value{
a list with two attributes:
\item{Est.prop}{a numeric matrix of cell type proportions estimated by CSsingle through maximizing the Spearman's correlation between the inferred and real bulk gene expression data (add up to 1)}
\item{Est.propS}{a ranked list of estimated cell type proportions}
\item{corr}{a numeric vector of Spearman's correlation coefficients}
}
\description{
This function is to combine CSsingle with each signature matrix to get estimates of the cell type proportions. The signature matrices were ordered in descending order based on the Spearman's correlation between their inferred and real bulk gene expression data.
}
\author{
Wenjun Shen
}
