% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessData.R
\name{preprocessData}
\alias{preprocessData}
\title{This function is to exclude cell types that are not interested from scRNA-seq data, remove sparse features and cells and trim bulk and scRNA-seq data to have the same features}
\usage{
preprocessData(
  bulk.eset,
  sc.eset,
  cellType,
  select.ct = NULL,
  cell.thres = 20,
  feature.thres = 0
)
}
\arguments{
\item{bulk.eset}{ExpressionSet object for bulk gene expression data}

\item{sc.eset}{SeuratObject for scRNA-seq data}

\item{cellType}{character, name of metadata column indicating cell type}

\item{select.ct}{character vector of cell types included, default is NULL, which means we include all cell types in the scRNA-seq data}

\item{cell.thres}{include features detected in at least this many cells}

\item{feature.thres}{include cells where at least this many features are detected}
}
\value{
a list with two attributes:
\item{bulk.eset}{bulk.eset}
\item{sc.eset}{SeuratObject for scRNA-seq data}
}
\description{
This function is to exclude cell types that are not interested from scRNA-seq data, remove sparse features and cells and trim bulk and scRNA-seq data to have the same features
}
\author{
Wenjun Shen
}
