% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualityControl.R
\name{qualityControl}
\alias{qualityControl}
\title{This function is to remove sparse features and cells from scRNA-seq data}
\usage{
qualityControl(sc.eset, cell.thres = 20, feature.thres = 0)
}
\arguments{
\item{sc.eset}{SeuratObject for scRNA-seq data}

\item{cell.thres}{include features detected in at least this many cells}

\item{feature.thres}{include cells where at least this many features are detected}
}
\value{
SeuratObject for scRNA-seq data after quality control
}
\description{
This function is to remove sparse features and cells from scRNA-seq data
}
\author{
Wenjun Shen
}
